% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_chi_squared_test.R
\name{bayesian_chi_squared_test}
\alias{bayesian_chi_squared_test}
\title{Bayesian Chi-squared test}
\usage{
bayesian_chi_squared_test(
  y,
  distribution_fun,
  params,
  data_type = "continuous",
  censored = NULL,
  chisq_quantile = 0.95
)
}
\arguments{
\item{y}{Independent and identically distributed samples drawn from a given distribution}

\item{distribution_fun}{The cumulative distribution function of the model posterior.}

\item{params}{Randomly sampled draws of from the parameters posterior distribution.}

\item{data_type}{There are three choices of data_type; "continuous", "discrete", and "censored".}

\item{censored}{A vector the same length as y indicating whether the data is censored.}

\item{chisq_quantile}{Quantile of the Chi-squared distribution to be used in the test.}
}
\value{
The proportion of R^b values that exceed the specified critical value from
their known Chi-squared reference distribution.
}
\description{
Bayesian Chi-squared test
}
