% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foocafeReliability-package.R
\docType{package}
\name{foocafeReliability-package}
\alias{foocafeReliability}
\alias{foocafeReliability-package}
\title{foocafeReliability: Bayesian Reliability Tools and Examples}
\description{
Tools and examples for use with the Bayesian reliability session at Foocafe in Malmö.
  Examples are taken from the book 'Bayesian Reliability' by Hamada, Wilson, Reese et al.
  Another example comes from the article 'Modelling Cesored Time-to-Event Data Using Pyro, an
  Open Source Probabilistic Programming Language'.
  (https://eng.uber.com/modeling-censored-time-to-event-data-using-pyro/)
}
\author{
\strong{Maintainer}: Andrew Segar \email{sesegaran@tetrapak.com}

}
\keyword{internal}
