% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcd_projector_failures}
\alias{lcd_projector_failures}
\title{Failure times (in projection hours) for 31 LCD projector lamps}
\format{\describe{A data frame with 31 rows and two columns
\item{lcd_model}{}
\item{projection_hours}{}
}}
\usage{
lcd_projector_failures
}
\description{
In business and educational settings, computer presentations use liquid
crystal display (LCD) projectors. The most common failure mode of these
projectors is the failure of the lamp. Many manufacturers include the
"expected" lamp life in their technical specification documents, and one
manufacturer claims that users can expect 1,500 hours of projection time
from each lamp used under "normal operating conditions". To test this claim,
a large private university placed identical lamps in three projector models
for a total of 31 projectors. The university staff recorded the number of
projection hours (as measured by the projector) when each lamp burned out.
This dataset are the results from that experiment.
}
\keyword{datasets}
