% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prowler_bearing_failures}
\alias{prowler_bearing_failures}
\title{Failure times (in operating hours) for bearings in 66 prowler attack
aircraft.}
\format{\describe{A data frame with 66 rows and three columns
\item{prowler_aircraft_id}{}
\item{operating_hours}{}
\item{right_censored}{}
}}
\usage{
prowler_bearing_failures
}
\description{
Prowler bearing failure times
}
\keyword{datasets}
