% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supercomputer_failures}
\alias{supercomputer_failures}
\title{Monthly number of failures for 47 supercomputer components}
\format{\describe{A data frame with 47 rows and two columns
\item{shared_memory_processor_id}{}
\item{failure_count}{}
}}
\usage{
supercomputer_failures
}
\description{
A dataset containing the monthly number of failures of the Los Alamos
National Laboratory Blue Mountain supercomputer components (shared memory
processors or SMPs).  The supercomputer consists of 47 "identical" SMPs this
dataset contains their montly number of failures for the first month of
operation.
}
\keyword{datasets}
